Move account operations to AccountRepository
