+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/sitepicker/SitePickerRepository.kt
@@ -29,16 +29,11 @@ import kotlin.coroutines.resume
 class SitePickerRepository @Inject constructor(
     private val siteStore: SiteStore,
     private val dispatcher: Dispatcher,
-    private val accountStore: AccountStore,
     private val wooCommerceStore: WooCommerceStore
 ) {
     suspend fun getSites() = withContext(Dispatchers.IO) { siteStore.sites }
 
     fun getSiteBySiteUrl(url: String) = SiteUtils.getSiteByMatchingUrl(siteStore, url)
-
-    fun getUserAccount() = accountStore.account
-
-    fun isUserLoggedIn() = accountStore.hasAccessToken()
 
     suspend fun fetchWooCommerceSites() = wooCommerceStore.fetchWooCommerceSites()
 
@@ -62,36 +57,6 @@ class SitePickerRepository @Inject constructor(
     suspend fun fetchSiteProductSettings(site: SiteModel) = wooCommerceStore.fetchSiteProductSettings(site)
 
     suspend fun verifySiteWooAPIVersion(site: SiteModel) = wooCommerceStore.fetchSupportedApiVersion(site)
-
-    suspend fun logout(): Boolean = suspendCancellableCoroutine { continuation ->
-        val listener = object : Any() {
-            @Suppress("unused")
-            @Subscribe(threadMode = ThreadMode.MAIN)
-            fun onAccountChanged(event: OnAccountChanged) {
-                if (event.causeOfChange == AccountAction.SIGN_OUT) {
-                    dispatcher.unregister(this)
-                    if (!continuation.isActive) return
-
-                    if (event.isError) {
-                        WooLog.e(
-                            WooLog.T.SITE_PICKER,
-                            "Account error [type = ${event.causeOfChange}] : " +
-                                "${event.error.type} > ${event.error.message}"
-                        )
-                        continuation.resume(false)
-                    } else if (!isUserLoggedIn()) {
-                        continuation.resume(true)
-                    }
-                }
-            }
-        }
-        dispatcher.dispatch(AccountActionBuilder.newSignOutAction())
-        dispatcher.dispatch(SiteActionBuilder.newRemoveWpcomAndJetpackSitesAction())
-
-        continuation.invokeOnCancellation {
-            dispatcher.unregister(listener)
-        }
-    }
 
     suspend fun fetchSiteInfo(siteAddress: String) =
         suspendCancellableCoroutine<Result<ConnectSiteInfoPayload>> { continuation ->
